Refactor LambdaLift
Break out the dependency collection into a separate class.
That makes it more re-usable and makes LambdaLift itself
easier to understand.
